+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
+## [Unreleased]
+
+Features:
+
+- add unit and e2e test for Jetton Wallet And Minter Wrappers
+
 ## [0.0.3] - 2023-04-22
 
 ### Fixed
 
-  - Update jetton minter content interface
+- Update jetton minter content interface
 
 ## [0.0.2] - 2023-04-22
 
@@ -12,7 +18,7 @@
 
 Features:
 
-  - add Jetton Wallet And Minter Wrappers
+- add Jetton Wallet And Minter Wrappers
 
 ## [0.0.1] - 2023-04-22
 
@@ -20,4 +26,4 @@ Features:
 
 Features:
 
-  - add onchain metadata and request libs
+- add onchain metadata and request libs